Add Particle System
You can do an experiment on this, but I suggest that we shouldn't implement this yet until I updated a stable version of this folder.

An html is provided on this folder to you guys to see the behaviour
testfile.js is where I've tested the particle system. you could check the configuration of that file as we well on the others